+++ CHANGELOG.md
@@ -4,10 +4,13 @@ Change Log
 Version 2.1.0 *(In Development)*
 --------------------------------
 
- * Added `ActionBarSherlock.Activity` and `ActionBarSherlock.ListActivity` for
-   extension by implementing activities, the latter of which is deprecated.
-   This affords a much tighter integration and allows for the use of other new
-   features listed below.
+**WARNING**: The [Android Compatibility Library (v4)][1] is now required.
+
+
+ * Added `ActionBarSherlock.Activity`, `ActionBarSherlock.ListActivity`,
+   and `ActionBarSherlock.FragmentActivity` for extension by implementing
+   activities, the latter of which is deprecated. This affords a much tighter
+   integration and allows for the use of other new features listed below.
  * New API method: `menu(int)` allows for the inflation of menu XMLs from a
    resource. For the non-native implementation, the XML is inflated to a custom
    Menu which can then be applied appropriately to the third-party action bar.
@@ -20,6 +23,9 @@ Version 2.1.0 *(In Development)*
    application icon/home button and title and show a larger logo representing
    the application. Third-party action bar handlers should implement
    `ActionBarSherlock.LogoHandler` for this functionality.
+ * New API method: `layout(Fragment)` will use the fragment argument as the
+   content to the activity. *This feature requires that your activity extends
+   from `ActionBarSherlock.FragmentActivity`*.
 
 
 Version 2.0.1 *(2011-03-11)*
@@ -43,3 +49,9 @@ Complete rewrite!
 Version 1.0.0 *(2011-03-07)*
 ----------------------------
 Initial release.
+
+
+
+
+
+ [1]: http://android-developers.blogspot.com/2011/03/fragments-for-all.html